Isolate dashboard types
