+++ src/models/dashboard/creatable-dashboard-tile.ts
@@ -1,0 +1,4 @@
+import { PartialProps } from '../../functions/utils';
+import { DashboardTile } from './dashboard-tile';
+
+export type CreatableDashboardTile = PartialProps<Omit<DashboardTile, 'id'>, 'rendererOptions'>;
