+++ src/models/dashboard/dashboard-live-update.ts
@@ -1,0 +1,1 @@
+export type DashboardLiveUpdate = { enabled: true; interval: number } | { enabled: false; interval?: null };
